% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labchecks.R
\name{check_commits}
\alias{check_commits}
\title{Check whether the user has at least three commits}
\usage{
check_commits(notebook)
}
\arguments{
\item{notebook}{A character indicating the path or list of paths to MDS lab
files (either.ipynb or .Rmd). If left blank, the function will recursively
search for all labs in the working directory based on the file extension.}
}
\value{
A logical which indicates whether the repo has 3 commits or not
}
\description{
Check whether the user has at least three commits
}
\examples{
\dontrun{
check_commits()
}
}
