% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labchecks.R
\name{check_lat_version}
\alias{check_lat_version}
\title{Check whether the user has pushed the latest version in his/her repository}
\usage{
check_lat_version(notebook)
}
\arguments{
\item{notebook}{A character indicating the path or list of paths to MDS lab
files (either.ipynb or .Rmd). If left blank, the function will recursively
search for all labs in the working directory based on the file extension.}
}
\value{
A logical which indicates whether the last pushed version is latest
or not
}
\description{
Check whether the user has pushed the latest version in his/her repository
}
\examples{
\dontrun{
check_lat_version()
}
}
