% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mechanics.R
\name{check_mechanics}
\alias{check_mechanics}
\title{Check MDS Lab Mechanics}
\usage{
check_mechanics(
  notebook = NULL,
  repo = ".",
  branch = usethis::git_branch_default()
)
}
\arguments{
\item{notebook}{A character indicating the path or list of paths to MDS lab
files (either.ipynb or .Rmd). If left blank, the function will recursively
search for all labs in the working directory based on the file extension.}

\item{repo}{A character path to the repo to check. Defaults to the working
directory.}

\item{branch}{A string specifying the name of the default branch used for
grading.}
}
\value{
The function prints the results of the mechanics checks to screen.
Silently returns TRUE if all the checks are passed.
}
\description{
Perform mechanics checks on a MDS Lab. This function checks that you have a
Github repo link, that you have pushed your latest commit, and that you have
at least three commit messages authored by you in your history.
}
\examples{
\dontrun{
check_mechanics()
}
}
