% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labchecks.R
\name{check_repo_link}
\alias{check_repo_link}
\title{Check whether the user has included the github repo link in his/her repository}
\usage{
check_repo_link(notebook)
}
\arguments{
\item{notebook}{A character indicating the path or list of paths to MDS lab
files (either.ipynb or .Rmd). If left blank, the function will recursively
search for all labs in the working directory based on the file extension.}
}
\value{
A logical which indicates whether the repo link exists or not
}
\description{
Check whether the user has included the github repo link in his/her repository
}
\examples{
\dontrun{
check_repo_link()
}
}
