% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countpoints.R
\name{count_points}
\alias{count_points}
\title{Tally Available Points in Lab}
\usage{
count_points(notebook = NULL, margins = TRUE)
}
\arguments{
\item{notebook}{A character indicating the path or list of paths to MDS lab
files (either.ipynb or .Rmd). If left blank, the function will recursively
search for all labs in the working directory based on the file extension.}

\item{margins}{A scalar logical which indicates whether to add a row for the
total number of points. Defaults to TRUE.}
}
\value{
A dataframe indicating total optional and required number of points.
}
\description{
Tally Available Points in Lab
}
\examples{
# R markdown
notebook <- system.file("extdata", "dummylab.Rmd", package = "labzenr")
count_points(notebook)

# Python notebook
notebook <- system.file("extdata", "dummylab.ipynb", package = "labzenr")
count_points(notebook)
}
