% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_assignment.R
\name{find_assignment}
\alias{find_assignment}
\title{Find a UBC MDS Lab Assignment}
\usage{
find_assignment()
}
\value{
The relative filepath to the lab found. Returns NULL if none found.
}
\description{
A helper function to locate the a lab file based on its extension being Rmd
ipynb. The utility will search recursively up the directory. If multiple
candidate files are found, the user will be prompted to select which file
they wish.
}
\examples{
# Navigate to a UBC lab repo and type
find_assignment()
}
