% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_lab}
\alias{parse_lab}
\title{Parse MDS lab files to return the markdown content}
\usage{
parse_lab(notebook = NULL)
}
\arguments{
\item{notebook}{A character indicating the path or list of paths to MDS lab
files (either.ipynb or .Rmd). If left blank, the function will recursively
search for all labs in the working directory based on the file extension.}
}
\value{
A character vector where each element is the content of one markdown cell.
}
\description{
Parse MDS lab files to return the markdown content
}
\examples{
\dontrun{
parse_lab()
}
}
