% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_signatures.R
\name{signature_add}
\alias{signature_add}
\title{Add a Git Signature to the User Cache}
\usage{
signature_add(x, is_regexp = TRUE, is_user = TRUE)
}
\arguments{
\item{x}{A character string.}

\item{is_regexp}{A logical}

\item{is_user}{A logical.}
}
\description{
Add a Git Signature to the User Cache
}
\examples{
\dontrun{
# adds a regular expression that matches a student username pattern
signature_add("John S(\\\\.|mith)", is_regexp = TRUE, is_user = TRUE)
}
}
