% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_signatures.R
\name{signature_clear}
\alias{signature_clear}
\title{Clear the labzen Cache of Known Git Signatures}
\usage{
signature_clear()
}
\description{
Clear the labzen Cache of Known Git Signatures
}
\examples{
\dontrun{
signatures_clear()
}

}
