% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_signatures.R
\name{signature_ls}
\alias{signature_ls}
\title{List Known Git Signatures from the User Cache}
\usage{
signature_ls(repo = ".")
}
\arguments{
\item{repo}{A path to a github repository. Defaults to the working
directory.}
}
\value{
A tibble containing three columns:
\enumerate{
\item The the \code{signature} column contains character strings of the
exact Git signatures (e.g. "John Smith \href{mailto:jsmith@example.com}{jsmith@example.com}") or
regular expressions (e.g. "^John S(mith|\\.)").
\item The \code{is_user} column indicates whether the signature/pattern belongs
to the student (FALSE would denote a Git user of MDS faculty).
\item The \code{is_regexp} column denotes if the signature column is a fixed
literal or a regular expression.
}
}
\description{
List Known Git Signatures from the User Cache
}
\examples{
# navigate to a Git repo and run:
signature_ls()
}
