% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_signatures.R
\name{signature_student}
\alias{signature_student}
\title{Return a List of Known Student Signatures}
\usage{
signature_student(repo = ".")
}
\arguments{
\item{repo}{A path to a github repository. Defaults to the working
directory.}
}
\value{
A logical indicating whether there are candidate Git signatures
that are not in the labzenr signature cache.
}
\description{
Return a List of Known Student Signatures
}
