% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_signatures.R
\name{signature_update}
\alias{signature_update}
\title{Interactively Update}
\usage{
signature_update(repo = ".", max_commits = 100L)
}
\arguments{
\item{repo}{A path to a github repository. Defaults to the working
directory.}

\item{max_commits}{An integer indicating the maximum commits to search in
in the git log.}
}
\description{
This function is intended to be run interactively from within an R session.
}
\examples{
\dontrun{
# navigate to a lab directory and run:
signature_update()
}
}
