% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_commits.R
\name{check_commits}
\alias{check_commits}
\title{Check whether the user has at least three commits}
\usage{
check_commits(...)
}
\arguments{
\item{...}{Arguments passed to \code{gert}, such as the \code{repo} name or \code{ref}.}
}
\value{
A logical which indicates whether the repo has 3 commits or not
}
\description{
Check whether the user has at least three commits
}
\examples{
\dontrun{
myrepo <- "/Users/sukhdeepkaur/MDS/Block5/lab/DSCI_563_lab1_sukh2929"
check_commits(repo = myrepo)
check_commits()
}

}
