% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_lat_version.R
\name{check_lat_version}
\alias{check_lat_version}
\title{Check whether the user has pushed the latest version in his/her repository}
\usage{
check_lat_version(repo = ".")
}
\arguments{
\item{repo}{A character path specifying the repo to check.}
}
\value{
A logical which indicates whether the last pushed version is latest
or not
}
\description{
Check whether the user has pushed the latest version in his/her repository
}
\examples{
\dontrun{
check_lat_version()
}
}
