% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labchecks.R
\name{check_mechanics}
\alias{check_mechanics}
\title{Performs Mechanics Checks on a MDS Lab This function check that you have a
Github repo link, that you have pushed your latest commit, and that you have
at least three commit messages authored by you in your history.}
\usage{
check_mechanics(
  notebook = NULL,
  pattern = NULL,
  fixed = TRUE,
  repo = ".",
  branch = usethis::git_branch_default()
)
}
\arguments{
\item{notebook}{A character indicating the path or list of paths to MDS lab
files (either.ipynb or .Rmd). If left blank, the function will recursively
search for all labs in the working directory based on the file extension.}

\item{pattern}{A character string indicating the a regular string (or
character string for \code{fixed = TRUE}) to be matched in the commit author
field.}

\item{fixed}{A logical indicating whether the pattern should be matched as
a literal (if TRUE) or a regular expression (if FALSE). Defaults to TRUE.}

\item{repo}{A character path to the repo to check. Defaults to the working
directory.}

\item{branch}{A string specifying the name of the default branch used for
grading.}
}
\value{
The function prints the results of the mechanics checks to screen.
Silently returns TRUE if all the checks are passed.
}
\description{
Performs Mechanics Checks on a MDS Lab This function check that you have a
Github repo link, that you have pushed your latest commit, and that you have
at least three commit messages authored by you in your history.
}
\examples{
\dontrun{
check_mechanics()
}
}
