% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labchecks.R
\name{check_repo_link}
\alias{check_repo_link}
\title{Check whether the user has included the github repo link in his/her
repository}
\usage{
check_repo_link(notebook = NULL)
}
\arguments{
\item{notebook}{A character indicating the path or list of paths to MDS lab
files (either.ipynb or .Rmd). If left blank, the function will recursively
search for all labs in the working directory based on the file extension.}
}
\value{
A logical which indicates whether the repo link exists or not
}
\description{
Check whether the user has included the github repo link in his/her
repository
}
\examples{
\dontrun{
# Python notebook example should work as it DOES not have a link
notebook <- system.file("extdata", "dummylab.ipynb", package = "labzenr")
check_repo_link(notebook)

# R markdown example should fail as it does NOT have a link
notebook <- system.file("extdata", "dummylab.Rmd", package = "labzenr")
check_repo_link(notebook)
}
}
